+++ CHANGELOG.md
@@ -13,5 +13,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Support to predict Appier ad
 - Allow to configure GDPR programmatically
 - Allow getting adapter version and network sdk version via `AppierAdapterConfiguration`